+++ src/ti/collection.c
@@ -128,6 +128,39 @@ void ti_collection_drop(ti_collection_t * collection)
         log_critical(EX_MEMORY_S);
 }
 
+int ti_collection_to_pk(ti_collection_t * collection, msgpack_packer * pk)
+{
+    return -(
+        msgpack_pack_map(pk, 8) ||
+
+        mp_pack_str(pk, "collection_id") ||
+        msgpack_pack_uint64(pk, collection->id) ||
+
+        mp_pack_str(pk, "name") ||
+        mp_pack_strn(pk, collection->name->data, collection->name->n) ||
+
+        mp_pack_str(pk, "created_at") ||
+        msgpack_pack_uint64(pk, collection->created_at) ||
+
+        mp_pack_str(pk, "things") ||
+        msgpack_pack_uint64(pk, collection->things->n + collection->gc->n) ||
+
+        mp_pack_str(pk, "time_zone") ||
+        mp_pack_strn(pk, collection->tz->name, collection->tz->n) ||
+
+        mp_pack_str(pk, "default_deep") ||
+        msgpack_pack_uint64(pk, collection->deep) ||
+
+        mp_pack_str(pk, "next_free_id") ||
+        msgpack_pack_uint64(pk, collection->next_free_id) ||
+
+        mp_pack_str(pk, "commit_history") ||
+        (collection->commits
+                ? msgpack_pack_uint32(pk, collection->commits->n)
+                : mp_pack_str(pk, "disabled"))
+    );
+}
+
 _Bool ti_collection_name_check(const char * name, size_t n, ex_t * e)
 {
     if (n < ti_collection_min_name || n >= ti_collection_max_name)